Fieed the problems with SPM packages not building problem, turned out to be a single "open class var" ( invalid ) declaration that was causing all the shit. I still need to work out the finer details of how page managemnt and page lists will worl. Will do that tomorrow.

+++ MedusaStorage/Sources/MedusaStorage/MedusaStorage.c
@@ -95,7 +95,7 @@ int16_t readInteger16(void* buffer,long offset)
     return(*from);
     }
     
-double readFloat(void* buffer,long offset)
+double readFloat64(void* buffer,long offset)
     {
     double* from = (double*)(((char*)buffer) + offset);
     return(*from);
